+++ flatlaf-core/src/main/java/com/formdev/flatlaf/ui/FlatUIUtils.java
@@ -38,6 +38,7 @@ import javax.swing.JComponent;
 import javax.swing.LookAndFeel;
 import javax.swing.UIManager;
 import javax.swing.plaf.ColorUIResource;
+import com.formdev.flatlaf.util.DerivedColor;
 import com.formdev.flatlaf.util.JavaCompatibility;
 import com.formdev.flatlaf.util.UIScale;
 
@@ -102,6 +103,12 @@ public class FlatUIUtils
 		g.setRenderingHint( RenderingHints.KEY_ANTIALIASING, RenderingHints.VALUE_ANTIALIAS_ON );
 		g.setRenderingHint( RenderingHints.KEY_STROKE_CONTROL,
 			MAC_USE_QUARTZ ? RenderingHints.VALUE_STROKE_PURE : RenderingHints.VALUE_STROKE_NORMALIZE );
+	}
+
+	public static void setColor( Graphics g, Color color, Color baseColor ) {
+		if( color instanceof DerivedColor )
+			color = ((DerivedColor)color).derive( baseColor );
+		g.setColor( color );
 	}
 
 	/**